// src/components/CollectionFilters/interfaces/props.tsx
import {
  CollectionFilterMetadata,
  CollectionMetadataFilter,
} from "../../../views/collections"

export interface CollectionFiltersProps {
  showActiveFilters: boolean
  selectedFilters: { [key: string]: string | null } | undefined
  addSelectedFilter: (
    filterName: string,
    selectedFilter: {
      value: any
      label: string
    }
  ) => void
  filtersFromMetadata: CollectionMetadataFilter[]
  resetFilters: (event: React.MouseEvent<HTMLButtonElement>) => void
  toggleActiveFilters: (event: React.MouseEvent<HTMLButtonElement>) => void
  filtersData: CollectionFilterMetadata[]
}